+++ src/Api/Controllers/LockersController.cs
@@ -1,7 +1,7 @@
 ﻿using Application.Common.Models;
 using Application.Common.Models.Dtos.Physical;
 using Application.Lockers.Commands.AddLocker;
-using Application.Lockers.Commands.RemoveLocker;
+using Application.Lockers.Commands.DisableLocker;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Api.Controllers;
@@ -16,7 +16,7 @@ public class LockersController : ApiControllerBase
     }
 
     [HttpPost("remove")]
-    public async Task<ActionResult<Result<LockerDto>>> RemoveLocker([FromBody] RemoveLockerCommand command)
+    public async Task<ActionResult<Result<LockerDto>>> RemoveLocker([FromBody] DisableLockerCommand command)
     {
         var result = await Mediator.Send(command);
         return Ok(Result<LockerDto>.Succeed(result));
